+++ CHANGELOG.rst
@@ -61,7 +61,7 @@ Humble v1.4.2 (2023-12-29)
 Humble v1.4.1 (2023-12-15)
 --------------------------
 * Removes the ``base_frame`` parameter from ``lbr_bringup``, ``lbr_description``, ``lbr_fri_ros2``, ``lbr_ros2_control``
-* Updates RViZ default config in ``lbr_moveit_config``
+* Updates RViz default config in ``lbr_moveit_config``
 * Refers to https://github.com/lbr-stack/lbr_fri_ros2_stack/pull/144
 
 Humble v1.4.0 (2023-12-08)